% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParams.R
\name{checkerList2vec}
\alias{checkerList2vec}
\title{Check if the provided parameter list is genuine}
\usage{
checkerList2vec(PARAMS_LIST, DIM_EXT, NYB, NYA, N_GRADES, N_EXAMS)
}
\arguments{
\item{PARAMS_LIST}{Parameter list. Structure to be documented. See the example.}

\item{DIM_EXT}{number of external covariates in the competing risk model.}

\item{NYB}{number of years in the non-graduatable state. Needed for determining how many time-related intercepts in the competing risk model.}

\item{NYA}{number of years in the graduatable state. Needed for determining how many time-related intercepts in the competing risk model.}

\item{N_GRADES}{number of grades modelled.}

\item{N_EXAMS}{number of exams.}
}
\description{
Check if the provided parameter list is genuine
}
