% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cr_conditional}
\alias{cr_conditional}
\title{Evaluate Outcome Likelihood}
\usage{
cr_conditional(
  THETA,
  EXTCOVARIATES,
  EXAMS_GRADES,
  EXAMS_DAYS,
  EXAMS_OBSFLAG,
  EXAMS_SET,
  OUTCOME,
  YEAR,
  N_GRADES,
  N_EXAMS,
  NYB,
  NYA,
  YEAR_LAST_EXAM,
  ABILITY,
  SPEED,
  GRFLAG = TRUE
)
}
\arguments{
\item{OUTCOME}{\code{1} for dropout, \code{2} for transfer, \code{3} for graduation. \code{0} if no outcome is observed.}

\item{NYB}{Total number of years in the non-graduatable regime. Needed for determining how many time-related intercepts.}

\item{NYA}{Total number of years in the graduatable regime. Needed for determining how many time-related intercepts.}

\item{YEAR_LAST_EXAM}{Year at which the all exams are completed for the first time.}

\item{GRFLAG}{\code{TRUE} to compute the gradient.}

\item{YEAR_FIRST}{First year to evaluate.}

\item{YEAR_LAST}{Last year to evaluate.}

\item{THETA_CR}{Portion of the parameter vector related to the competing risk model}

\item{COVARIATES}{The first 2 values refers to ability and speed respectively. Remaining values are external predictors.}
}
\description{
Evaluate Outcome Likelihood
}
