% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParams.R
\name{extract_params}
\alias{extract_params}
\title{Extract parameters of interest from theta vector}
\usage{
extract_params(
  THETA,
  DIM_EXT,
  NYB,
  NYA,
  N_GRADES,
  N_EXAMS,
  PAR_TYPE = NA,
  OUTCOME = NA,
  EXAM = NA
)
}
\arguments{
\item{THETA}{parameter vector}

\item{DIM_EXT}{number of external covariates in the competing risk model.}

\item{NYB}{number of years in the non-graduatable state. Needed for determining how many time-related intercepts in the competing risk model.}

\item{NYA}{number of years in the graduatable state. Needed for determining how many time-related intercepts in the competing risk model.}

\item{N_GRADES}{number of grade-block modelled.}

\item{N_EXAMS}{number of exams.}

\item{PAR_TYPE}{1 for regression coeff, 0 for intercepts.}

\item{OUTCOME}{outcome of interest. Possible values: \code{d}, \code{t} and \code{g}.}

\item{EXAM}{exam of interest. Posible values in \code{1:N_EXAMS}.}
}
\value{
It returns a vector containing the parameters of interests if either \code{OUTCOME} or \code{EXAM} have been specified.
It returns a complete list of the parameters if both \code{OUTCOME} and \code{EXAM} are set to \code{NA}.
}
\description{
Extract parameters of interest from theta vector
}
