% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_params_cr}
\alias{extract_params_cr}
\title{Extract parameters related to the competing risk model}
\usage{
extract_params_cr(THETA_CR, DIM_EXT, NYB, NYA, OPTION)
}
\description{
Extract parameters related to the competing risk model
}
