% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_params_idx_cr}
\alias{extract_params_idx_cr}
\title{Extract parameters indexes related to the competing risk model}
\usage{
extract_params_idx_cr(THETA_CR, DIM_EXT, NYB, NYA, OPTION)
}
\arguments{
\item{THETA_CR}{Parameter vector related to the competing risk model}

\item{DIM_EXT}{Number of external covariates in the competing risk model.}

\item{NYB}{Number of years in the non-graduatable state. Needed for
determining how many time-related intercepts in the competing risk model.}

\item{NYA}{Number of years in the graduatable state.
Needed for determining how many time-related intercepts in the competing risk model.}

\item{OPTION}{It selects the parameters of interest.
\code{1} for beta_d, \code{2} for beta_t, \code{3} for beta_g, \code{4} for beta0_d,
\code{5} for beta0_t, \code{6} for beta0_g.}
}
\value{
It return a vector with two values representing the starting
index and the length of the segment of THETA_CR corresponding to
the parameter of interest
}
\description{
Extract parameters indexes related to the competing risk model
}
