% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_params_irt}
\alias{extract_params_irt}
\title{Extract parameters related to the IRT model}
\usage{
extract_params_irt(THETA_IRT, N_GRADES, N_EXAMS, OPTION, EXAM)
}
\description{
Extract parameters related to the IRT model
}
