% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{full_model}
\alias{full_model}
\title{Full model}
\usage{
full_model(
  THETA,
  EXTCOVARIATES,
  EXAMS_GRADES,
  EXAMS_DAYS,
  EXAMS_OBSFLAG,
  EXAMS_SET,
  OUTCOME,
  YEAR,
  YEAR_LAST_EXAM,
  N_GRADES,
  N_EXAMS,
  NYB,
  NYA,
  ABILITY,
  SPEED,
  GRFLAG = TRUE
)
}
\arguments{
\item{THETA}{Suitable parameter vector as provided by \link{paramsList2vec}().}

\item{EXTCOVARIATES}{External covariates.}

\item{EXAMS_GRADES}{Vector of grades.}

\item{EXAMS_DAYS}{Vector of times.}

\item{EXAMS_OBSFLAG}{Vector of booleans.\code{TRUE} elements represent observed exams. \code{FALSE} elements the unobserved ones.}

\item{EXAMS_SET}{Vector filled with booleans.\code{TRUE} elements represent exams in the study plan. \code{FALSE} elements non-relevant ones.}

\item{OUTCOME}{\code{1} for dropout, \code{2} for transfer, \code{3} for graduation. \code{0} if no outcome is observed.}

\item{YEAR}{Year of evaluation.}

\item{YEAR_LAST_EXAM}{Year at which the all exams are completed for the first time.}

\item{N_GRADES}{Number of grades modelled.}

\item{N_EXAMS}{Number of exams modelled}

\item{NYB}{Number of years in the non-graduatable state. Needed for determining how many time-related intercepts.}

\item{NYA}{Number of years in the graduatable state. Needed for determining how many time-related intercepts.}

\item{ABILITY}{Ability value.}

\item{SPEED}{Speed value.}

\item{GRFLAG}{\code{TRUE} to compute the gradient.}
}
\value{
It returns the value of the integrand function,
given the parameters and the data of a single observation.
}
\description{
Evaluate the full model log likelihood of a single observation
}
