% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gr_pGreaterGrades}
\alias{gr_pGreaterGrades}
\title{Evaluate the probability of grades greater or equal than the reference one}
\usage{
gr_pGreaterGrades(GRADE, EXAM, THETA_IRT, N_GRADES, N_EXAMS, ABILITY)
}
\arguments{
\item{GRADE}{Grade used as reference}

\item{EXAM}{Exam of interest}

\item{THETA_IRT}{Portion of the parameter vector related to the IRT model}

\item{N_GRADES}{Number of grades modelled.}

\item{N_EXAMS}{Number of exams.}

\item{ABILITY}{Ability value.}

\item{LOGFLAG}{Set TRUE to return log value.}
}
\value{
It returns the probability of obtaining grades higher than \code{GRADE} on exam \code{EXAM}.
}
\description{
Evaluate the probability of grades greater or equal than the reference one
}
