% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grl_examLik}
\alias{grl_examLik}
\title{Evaluate exam specific likelihood}
\usage{
grl_examLik(
  EXAM,
  GRADE,
  DAY,
  MAX_DAY,
  OBSFLAG,
  THETA_IRT,
  N_GRADES,
  N_EXAMS,
  ABILITY,
  SPEED,
  LOGFLAG = FALSE
)
}
\arguments{
\item{EXAM}{Exam of interest.}

\item{GRADE}{Grade of interest.}

\item{DAY}{Day of interest.}

\item{MAX_DAY}{Last day of observation.}

\item{OBSFLAG}{TRUE for observed, FALSE for not-observed.}

\item{THETA_IRT}{Portion of the parameter vector related to the IRT model}

\item{N_GRADES}{Number of grades modelled.}

\item{N_EXAMS}{Number of exams.}

\item{ABILITY}{ability value.}

\item{SPEED}{speed value.}

\item{LOGFLAG}{Set TRUE to return log value.}
}
\value{
It returns the probability of observing or not a specific
grade on a given exam before a given day conditioned on ability and speed.
}
\description{
Evaluate exam specific likelihood
}
