% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hazard}
\alias{hazard}
\title{Evaluate hazard function based on outcome and year}
\usage{
hazard(OUTCOME, YEAR, THETA_CR, COVARIATES, NYB, NYA)
}
\arguments{
\item{OUTCOME}{1 for dropout, 2 for transfer, 3 for graduation}

\item{YEAR}{Possible values 1:NYB in case of dropout/transfer; 1:NYA in case of graduation}

\item{THETA_CR}{Portion of the parameter vector related to the competing risk model}

\item{COVARIATES}{First 2 values for ability and speed. Remaining values are external covariates}

\item{NYB}{number of years in the non-graduatable state. Needed for determining how many time-related intercepts.}

\item{NYA}{number of years in the graduatable state. Needed for determining how many time-related intercepts.}
}
\value{
It returns the hazard probability of the specific outcome and year.
}
\description{
Evaluate hazard function based on outcome and year
}
