% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{irt_conditional}
\alias{irt_conditional}
\title{Evaluate the complete IRT log likelihood function}
\usage{
irt_conditional(
  THETA,
  EXAMS_GRADES,
  EXAMS_DAYS,
  EXAMS_OBSFLAG,
  EXAMS_SET,
  YEAR,
  N_GRADES,
  N_EXAMS,
  NYB,
  NYA,
  DIM_EXT,
  ABILITY,
  SPEED,
  GRFLAG = TRUE
)
}
\arguments{
\item{THETA}{Suitable parameter vector as provided by \link{paramsList2vec}().}

\item{EXAMS_GRADES}{Vector of grades.}

\item{EXAMS_DAYS}{Vector of times.}

\item{EXAMS_OBSFLAG}{Vector of booleans.\code{TRUE} elements represent observed exams. \code{FALSE} elements the unobserved ones.}

\item{EXAMS_SET}{Vector filled with booleans.\code{TRUE} elements represent exams in the study plan. \code{FALSE} elements non-relevant ones.}

\item{YEAR}{Year of evaluation.}

\item{N_GRADES}{Number of grades modelled.}

\item{N_EXAMS}{Number of exams modelled}

\item{NYB}{Number of years in the non-graduatable state. Needed for determining how many time-related intercepts.}

\item{NYA}{Number of years in the graduatable state. Needed for determining how many time-related intercepts.}

\item{DIM_EXT}{Number of external covariates in the CR conditional model.}

\item{ABILITY}{Ability value.}

\item{SPEED}{Speed value.}

\item{GRFLAG}{\code{TRUE} to compute the gradient.}
}
\value{
It returns a list with:
\itemize{
\item ll - The log-likelihood of the conditional IRT model.
\item gr - The gradient of the log-likelihood.
}
}
\description{
Evaluate the complete IRT log likelihood function
}
