% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lat_distr}
\alias{lat_distr}
\title{Latent distribution}
\usage{
lat_distr(ABILITY, SPEED, THETA, GRFLAG = TRUE)
}
\arguments{
\item{ABILITY}{ability value.}

\item{SPEED}{speed value.}

\item{THETA}{parameters vector. The last two elements are used as (L\link{2,1}, L\link{2,2}), where L is the lower triangular Cholesky of the latent covariance matrix.}

\item{GRFLAG}{\code{TRUE} to compute the gradient.}
}
\value{
It returns a list with:
\itemize{
\item ll - The log-likelihood of the latent distribution.
\item gr - The gradient of the log-likelihood.
}
}
\description{
Provides a wrapper for the cpp class \code{LAT_DISTR}, which computes
log density and gradient of the joint distribution of ability and speed.
}
