% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{latent_distr}
\alias{latent_distr}
\title{Joint density of speed and ability}
\usage{
latent_distr(ABILITY, SPEED, REPRHO, REPSIGMA, LOGFLAG = FALSE)
}
\arguments{
\item{ABILITY}{ability value.}

\item{SPEED}{speed value.}

\item{REPRHO}{atanh transformation of the latent correlation.}

\item{REPSIGMA}{log standard deviation of speed.}

\item{LOGFLAG}{\code{TRUE} for log density.}
}
\description{
Joint density of speed and ability
}
