% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{latent_distr2}
\alias{latent_distr2}
\title{Joint density of speed and ability}
\usage{
latent_distr2(ABILITY, SPEED, PAR1, PAR2, LOGFLAG = FALSE)
}
\arguments{
\item{ABILITY}{ability value.}

\item{SPEED}{speed value.}

\item{PAR1}{L\link{2,1}.}

\item{PAR2}{L\link{2,2}.}

\item{LOGFLAG}{\code{TRUE} for log density.}
}
\description{
Joint density of speed and ability
}
