% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalL.R
\name{marginal_likelihood}
\alias{marginal_likelihood}
\title{Approximate the marginal likelihood of a single observation}
\usage{
marginal_likelihood(
  THETA,
  EXTCOVARIATES,
  EXAMS_GRADES,
  EXAMS_DAYS,
  EXAMS_OBSFLAG,
  OUTCOME,
  YEAR,
  N_GRADES,
  N_EXAMS,
  NYB,
  NYA,
  YEAR_LAST_EXAM = 100,
  LOGFLAG = FALSE,
  CUBATURE_METHOD = "hcubature"
)
}
\arguments{
\item{THETA}{Suitable parameter vector as provided by \link{paramsList2vec}().}

\item{EXTCOVARIATES}{Vector of external covariates.}

\item{EXAMS_GRADES}{Vector of grades.}

\item{EXAMS_DAYS}{Vector of times.}

\item{EXAMS_OBSFLAG}{Vector of booleans.\code{TRUE} elements represent observed exams. \code{FALSE} elements the unobserved ones.}

\item{OUTCOME}{\code{1} for dropout, \code{2} for transfer, \code{3} for graduation. \code{0} if no outcome is observed.}

\item{YEAR}{Year of evaluation.}

\item{N_GRADES}{Number of grades modelled.}

\item{N_EXAMS}{Number of exams modelled}

\item{NYB}{Number of years in the non-graduatable state. Needed for determining how many time-related intercepts.}

\item{NYA}{Number of years in the graduatable state. Needed for determining how many time-related intercepts.}

\item{YEAR_LAST_EXAM}{Year at which the all exams are completed for the first time.}

\item{LOGFLAG}{Set TRUE to return log value.}

\item{CUBATURE_METHOD}{Argument to be passed to \link[cubature:cubintegrate]{cubature::cubintegrate}.}
}
\description{
Approximate the marginal likelihood of a single observation
}
