% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParams.R
\name{paramsVec2list}
\alias{paramsVec2list}
\title{Construct named list from parameter vector}
\usage{
paramsVec2list(THETA, DIM_EXT, NYB, NYA, N_GRADES, N_EXAMS, CONSTRFLAG = T)
}
\arguments{
\item{THETA}{Parameter vector. Structure to be documented.}

\item{DIM_EXT}{number of external covariates in the competing risk model.}

\item{NYB}{number of years in the non-graduatable state. Needed for determining how many time-related intercepts in the competing risk model.}

\item{NYA}{number of years in the graduatable state. Needed for determining how many time-related intercepts in the competing risk model.}

\item{N_GRADES}{number of grades modelled.}

\item{N_EXAMS}{number of exams.}

\item{CONSTRFLAG}{TRUE if input is defined in the constrained parameter space.}
}
\description{
Construct named list from parameter vector
}
