% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParams.R
\name{reparInt}
\alias{reparInt}
\title{Intercepts reparameterisation}
\usage{
reparInt(X, CON2UN = T)
}
\arguments{
\item{X}{Intercepts or unconstrained parameter values for grades low-to-high.}

\item{CON2UN}{TRUE if going from the constrained space to the unconstrained one.}
}
\value{
The numeric vector containing the constrained intercepts
or the unconstrained parameters.
}
\description{
It allows to go back and forth from constrained intercepts
to unconstrained parameters.
}
\examples{
set.seed(123)

# random unconstrained parameters
params <- rnorm(4)

# from unconstrained parameters to constrained intercepts
intercepts <- reparInt(params, CON2UN = FALSE)

# retrieve params
reparInt(intercepts, CON2UN = TRUE)

}
