% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParams.R
\name{reparThres}
\alias{reparThres}
\title{Threshold reparameterisation}
\usage{
reparThres(THRESHOLDS, UN2CON = T)
}
\arguments{
\item{THRESHOLDS}{Thresholds}

\item{UN2CON}{TRUE if going from the unconstrained space to the constrained one.}
}
\description{
Threshold reparameterisation
}
