% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData.R
\name{rngGrades}
\alias{rngGrades}
\title{Generate random grades}
\usage{
rngGrades(THETA_IRT, N_GRADES, N_EXAMS, ABILITY, SEED = 123)
}
\arguments{
\item{THETA_IRT}{Portion of the parameter vector related to the IRT model.}

\item{N_GRADES}{Number of grades modelled.}

\item{N_EXAMS}{Number of exams.}

\item{ABILITY}{Ability value.}

\item{SEED}{Random seed.}
}
\value{
It returns a vector of grades.
}
\description{
Generate random grades
}
