% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData.R
\name{rngTimes}
\alias{rngTimes}
\title{Generate random Times}
\usage{
rngTimes(THETA_IRT, N_GRADES, N_EXAMS, SPEED, SEED = 123)
}
\arguments{
\item{THETA_IRT}{Portion of the parameter vector related to the IRT model.}

\item{N_GRADES}{Number of grades modelled.}

\item{N_EXAMS}{Number of exams.}

\item{SPEED}{Speed value.}

\item{SEED}{Random seed.}
}
\value{
It returns a vector of times
}
\description{
Generate random Times
}
