% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalL.R
\name{sample_marginal_likelihood}
\alias{sample_marginal_likelihood}
\title{Approximate the marginal likelihood of the sample}
\usage{
sample_marginal_likelihood(
  THETA,
  EXTCOVARIATES,
  EXAMS_GRADES,
  EXAMS_DAYS,
  EXAMS_OBSFLAG,
  OUTCOME,
  YEAR,
  N_GRADES,
  N_EXAMS,
  NYB,
  NYA,
  YEAR_LAST_EXAM = 100,
  LOGFLAG = FALSE,
  CUBATURE_METHOD = "hcubature"
)
}
\arguments{
\item{THETA}{Suitable parameter vector as provided by \link{paramsList2vec}().}

\item{EXTCOVARIATES}{Matrix of external covariates with \code{n} rows.}

\item{EXAMS_GRADES}{List of \code{n} vectors of grades.}

\item{EXAMS_DAYS}{List of \code{n} vectors of times.}

\item{EXAMS_OBSFLAG}{VList of \code{n} vectors of booleans.\code{TRUE} elements represent observed exams. \code{FALSE} elements the unobserved ones.}

\item{OUTCOME}{Vector of \code{n} outcomes: \code{1} for dropout, \code{2} for transfer, \code{3} for graduation. \code{0} if no outcome is observed.}

\item{YEAR}{Vector of \code{n} years of evaluation.}

\item{N_GRADES}{Number of grades modelled.}

\item{N_EXAMS}{Number of exams modelled}

\item{NYB}{Number of years in the non-graduatable state. Needed for determining how many time-related intercepts.}

\item{NYA}{Number of years in the graduatable state. Needed for determining how many time-related intercepts.}

\item{YEAR_LAST_EXAM}{Year at which the all exams are completed for the first time.}

\item{LOGFLAG}{Set TRUE to return log value.}

\item{CUBATURE_METHOD}{Argument to be passed to \link[cubature:cubintegrate]{cubature::cubintegrate}.}
}
\description{
Approximate the marginal likelihood of the sample
}
