% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{survival}
\alias{survival}
\title{Evaluate survival function given a the range of years of interest}
\usage{
survival(YEAR_FIRST, YEAR_LAST, REGIMEFLAG, THETA_CR, COVARIATES, NYB, NYA)
}
\arguments{
\item{YEAR_FIRST}{First year to evaluate.}

\item{YEAR_LAST}{Last year to evaluate.}

\item{REGIMEFLAG}{\code{TRUE} if graduation is possible, \code{FALSE} otherwise.}

\item{THETA_CR}{Portion of the parameter vector related to the competing risk model}

\item{COVARIATES}{The first 2 values refers to ability and speed respectively. Remaining values are external predictors.}

\item{NYB}{Total number of years in the non-graduatable regime. Needed for determining how many time-related intercepts.}

\item{NYA}{Total number of years in the graduatable regime. Needed for determining how many time-related intercepts.}
}
\value{
It returns the hazard probability of the specific outcome and year.
}
\description{
Evaluate survival function given a the range of years of interest
}
