% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParams.R
\name{paramsVec2list}
\alias{paramsVec2list}
\title{Construct named list from parameter vector}
\usage{
paramsVec2list(THETA, DIM_EXT, NYB, NYA, N_GRADES, N_EXAMS)
}
\arguments{
\item{THETA}{Parameter vector. Structure to be documented.}

\item{DIM_EXT}{number of external covariates in the competing risk model.}

\item{NYB}{number of years in the non-graduatable state. Needed for determining how many time-related intercepts in the competing risk model.}

\item{NYA}{number of years in the graduatable state. Needed for determining how many time-related intercepts in the competing risk model.}

\item{N_GRADES}{number of grades modelled.}

\item{N_EXAMS}{number of exams.}
}
\value{
Given a proper unconstrained parameter vector, it returns a
list of the model parameters.
}
\description{
Construct named list from parameter vector
}
\examples{

# Use example from paramsList2vec() documentation
# to get a proper example of parameter vector
example(paramsList2vec)

# Retrieve the list
params_list <- paramsVec2list(
  THETA = theta,
  DIM_EXT = external_covariates,
  NYB = years_before,
  NYA = years_after,
  N_GRADES = grades,
  N_EXAMS = exams)



}
