% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParams.R
\name{paramsVec2list}
\alias{paramsVec2list}
\title{Construct named list from parameter vector}
\usage{
paramsVec2list(
  THETA,
  DIM_EXT,
  NYB,
  NYA,
  N_GRADES,
  N_EXAMS,
  LABS_EXAMS = NULL,
  LABS_GRADES = NULL,
  LABS_COVARIATES = NULL
)
}
\arguments{
\item{THETA}{Parameter vector. Structure to be documented.}

\item{DIM_EXT}{number of external covariates in the competing risk model.}

\item{NYB}{number of years in the non-graduatable state. Needed for determining how many time-related intercepts in the competing risk model.}

\item{NYA}{number of years in the graduatable state. Needed for determining how many time-related intercepts in the competing risk model.}

\item{N_GRADES}{number of grades modelled.}

\item{N_EXAMS}{number of exams.}

\item{LABS_EXAMS}{optional label for exams}

\item{LABS_GRADES}{optional label for grades}

\item{LABS_COVARIATES}{optional label for covariates}
}
\value{
Given a proper unconstrained parameter vector, it returns a
list of the model parameters.
}
\description{
Construct named list from parameter vector
}
\examples{

seed <- 123

## Setup dimensions
external_covariates <- 3L     # number of external covariates
years_before <- 6L            # number of possible years in the regime without graduation
years_after <- 2L             # number of possible years in the regime with graduation
grades <- 4L                  # number of grades
exams <- 10L                  # number of exams

dim <- 3 * (external_covariates+2) + 2 * (years_before) +
        years_after + 3 * exams + exams * grades + 2

set.seed(seed)
theta <- rnorm(dim)

## Retrieve the list
params_list <- paramsVec2list(
  THETA = theta,
  DIM_EXT = external_covariates,
  NYB = years_before,
  NYA = years_after,
  N_GRADES = grades,
  N_EXAMS = exams)



}
